+++ lib_gnome/NetCDFWindMover_c.h
@@ -65,7 +65,7 @@ public:
 	NetCDFWindMover_c (TMap *owner, char* name);
 	NetCDFWindMover_c () {}
 	virtual OSErr 		PrepareForModelRun(); 
-	virtual OSErr 		PrepareForModelStep(const Seconds&, const Seconds&, bool, int numLESets, int* LESetsSizesList); 
+	virtual OSErr 		PrepareForModelStep(const Seconds&, const Seconds&, bool, int numLESets, double* LESetsSizesList); 
 	virtual void 		ModelStepIsDone();
 	virtual WorldPoint3D       GetMove(const Seconds& model_time, Seconds timeStep,long setIndex,long leIndex,LERec *theLE,LETYPE leType);
 	virtual long 		GetVelocityIndex(WorldPoint p);
